Extract UnitOrganizationTree component
